backend/requirements.txt
# Conecta Plus - Backend Requirements

# FastAPI e Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Redis (cache e filas)
redis==5.0.1
celery==5.3.6

# Utilities
python-dateutil==2.8.2
httpx==0.26.0
aiofiles==23.2.1

# Monitoring
prometheus-client==0.19.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Development
black==24.1.1
isort==5.13.2
flake8==7.0.0
